requirements.txt
pandas
openpyxl
sentence-transformers
faiss-cpu
langchain
langchain-community
openai
tiktoken
